refactor: store 간 구별을 위해 store -> newsStore 로 파일명 변경
